feat(company): add company search functionality
implemented a new search endpoint in CompanyRestController to allow searching for companies by name. Added a corresponding method in CompanyService and CompanyRepository to handle the search query. In the front-end, enhanced company management by adding a search bar and dynamic table population for displaying search results. Improved input formatting for CNPJ and phone numbers in company-script.js and company.html.